dedup remaining parsers
